【背景】 codex 告警清零 【修改方案】 修改和添加了procfs的默认文件模式的宏定义,对进行位操作运算的有符号数进行强制类型转换
【影响】
对现有的产品编译不会有影响。

re #I4KVS7

+++ syscall/fs_syscall.c
@@ -307,7 +307,7 @@ int SysOpen(const char *path, int oflags, ...)
         goto ERROUT;
     }
 
-    if (oflags & O_CLOEXEC) {
+    if ((unsigned int)oflags & O_CLOEXEC) {
         SetCloexecFlag(procFd);
     }
 
@@ -1640,7 +1640,7 @@ int SysOpenat(int dirfd, const char *path, int oflags, ...)
         goto ERROUT;
     }
 
-    if (oflags & O_CLOEXEC) {
+    if ((unsigned int)oflags & O_CLOEXEC) {
         SetCloexecFlag(procFd);
     }
